delete image with movie_info_page.dart
