+++ lib/pages/play_trailer_page.dart
@@ -1,7 +1,6 @@
-import 'dart:async';
-
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:kinopoisk/widgets/app_bar_widget.dart';
 //import 'package:webview_flutter/webview_flutter.dart';
